+++ src/formatter/terminal.rs
@@ -66,7 +66,7 @@ impl Formatter for Terminal<'_> {
                 self.lang.config(),
                 self.source.as_bytes(),
                 None,
-                |injected| Some(Language::guess(injected, "").config()),
+                |injected| Some(Language::guess(Some(injected), "").config()),
             )
             .expect("failed to generate highlight events");